# openapi.yaml
openapi: 3.1.0
info:
  title: crypto-gateway
  version: 0.0.0
  description: |
    Runtime-agnostic crypto payment gateway with pluggable chain adapters
    (EVM / Tron / Solana), DB adapters (D1 / libSQL / Postgres), cache
    adapters, detection strategies, and webhook delivery.

    Deploys on Cloudflare Workers, Node/Bun, Deno, and Vercel Edge from one
    codebase.
servers:
  - url: http://localhost:8787
    description: Local Node dev
  - url: https://crypto-gateway-dev.workers.dev
    description: Dev Workers (placeholder — replace with your subdomain)
...
security:
  - MerchantApiKey: []
paths:
...
  /webhooks/alchemy:
    post:
      tags: [webhooks]
      summary: Ingest Alchemy Notify ADDRESS_ACTIVITY webhooks
      description: |
        HMAC-SHA256 of the raw body under the per-webhook signing key
        (looked up by `event.webhookId` in the encrypted-at-rest
        `alchemy_webhook_registry` table) must match `X-Alchemy-Signature`.
        Returns 200 fast and processes the payload in the background;
        Alchemy retries on non-2xx.
      security:
        - AlchemySignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: Acknowledged; ingest happens in background
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { type: boolean, const: true }
        "400": { description: Malformed JSON after a valid signature }
        "401": { description: Missing or bad signature }
        "404": { description: Alchemy ingest not enabled on this deployment }
        "429": { $ref: "#/components/responses/RateLimited" }
  /admin/merchants:
    post:
      tags: [admin]
      summary: Create a merchant
      description: |
        Returns the plaintext API key ONCE in the response. The hash is
        stored; the plaintext is unrecoverable afterward.
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 128 }
                webhookUrl: { type: string, format: uri }
                webhookSecret:
                  type: string
                  description: Optional hex-encoded 32-byte HMAC signing secret. Generated if omitted.
                  pattern: "^[0-9a-f]{64}$"
      responses:
        "201":
          description: Merchant created
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      name: { type: string }
                      webhookUrl: { type: [string, "null"] }
                      active: { type: boolean }
                      createdAt: { type: string, format: date-time }
                  apiKey:
                    type: string
                    description: Plaintext API key. Shown once; store it immediately.
                    pattern: "^sk_[0-9a-f]{64}$"
                  webhookSecret:
                    type: string
                    description: Present only if `webhookUrl` was provided.
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Admin surface not enabled (ADMIN_KEY not set) }
  /admin/bootstrap/alchemy-webhooks:
    post:
      tags: [admin]
      summary: Idempotently create Alchemy webhooks for the configured chains
      description: |
        For each requested chainId, checks whether an Alchemy ADDRESS_ACTIVITY
        webhook already exists with the target URL + network; creates it if
        not. Returns per-chain status plus the `signingKey` for any newly
        created webhook. Signing keys are persisted encrypted into
        `alchemy_webhook_registry` automatically; the `signingKey` field in
        the response is a one-shot safety net in case persistence raced.
        Requires `ALCHEMY_NOTIFY_TOKEN` env.
      security:
        - AdminKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                chainIds:
                  type: array
                  items: { type: integer }
                  description: Defaults to `ALCHEMY_CHAINS` env, then the mainnet set
                seedAddressByChainId:
                  type: object
                  additionalProperties: { type: string }
                  description: Per-chain placeholder address (Alchemy requires ≥1 at create). Defaults to zero-address.
              description: |
                The target webhook URL is NOT accepted in the body. It is sourced from
                `GATEWAY_PUBLIC_URL` env (base origin) plus the canonical path
                `/webhooks/alchemy`. Env-only so a leaked ADMIN_KEY cannot be
                used to redirect Alchemy's traffic to an arbitrary host.
      responses:
        "200":
          description: Bootstrap complete; per-chain status in body
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        chainId: { type: integer }
                        status:
                          type: string
                          enum: [created, existing, unsupported, failed]
                        webhookId: { type: string }
                        signingKey:
                          type: string
                          description: Only present when `status=created`. Automatically persisted to `alchemy_webhook_registry`; kept in the response as a one-shot safety net if `persisted=false`.
                        error: { type: string }
        "400":
          description: Missing ALCHEMY_NOTIFY_TOKEN or GATEWAY_PUBLIC_URL env
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Admin surface not enabled (ADMIN_KEY not set) }
  /admin/alchemy-webhooks/signing-keys:
    post:
      tags: [admin]
      summary: Register an Alchemy webhook's signing key manually
      description: |
        For operators who created a webhook through the Alchemy dashboard UI
        (not via bootstrap) or who need to push a key back in after a
        persistence failure. The signing key is encrypted via
        `SECRETS_ENCRYPTION_KEY` before being written to
        `alchemy_webhook_registry`. Re-running with the same chainId upserts
        (rotation after a dashboard delete+recreate).
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chainId, webhookId, signingKey, webhookUrl]
              properties:
                chainId: { type: integer }
                webhookId: { type: string, minLength: 1, maxLength: 128 }
                signingKey: { type: string, minLength: 1, maxLength: 256 }
                webhookUrl: { type: string, format: uri }
      responses:
        "201":
          description: Signing key persisted encrypted-at-rest
          content:
            application/json:
              schema:
                type: object
                properties:
                  registered:
                    type: object
                    properties:
                      chainId: { type: integer }
                      webhookId: { type: string }
                      webhookUrl: { type: string, format: uri }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Admin surface not enabled (ADMIN_KEY not set) }
  /admin/fee-wallets:
    post:
      tags: [admin]
      summary: Register a fee wallet + store its private key in the SignerStore
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chainId, address, label, privateKey, family]
              properties:
                chainId: { type: integer }
                address: { type: string }
                label: { type: string, maxLength: 64 }
                privateKey:
                  type: string
                  description: Hex or family-specific. Stored in the SignerStore at `{kind:"fee-wallet", family, label}`.
                family:
                  type: string
                  enum: [evm, tron, solana]
      responses:
        "201":
          description: Fee wallet registered
        "400":
          description: Invalid address OR family does not match the chain adapter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    MerchantApiKey:
      type: http
      scheme: bearer
      description: |
        Merchant API key returned once at `POST /admin/merchants`.
        Also accepted as `X-API-Key: <key>`.
    AdminKey:
      type: http
      scheme: bearer
      description: Shared admin key from `ADMIN_KEY` env. Also accepted as `X-Admin-Key`.
    AlchemySignature:
      type: apiKey
      in: header
      name: X-Alchemy-Signature
      description: HMAC-SHA256 hex of the raw body under the per-webhook signing key stored in `alchemy_webhook_registry` (looked up by `event.webhookId`).
    CronSecret:
      type: http
      scheme: bearer
      description: Shared secret from `CRON_SECRET` env.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Request body failed schema validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests in the current window
      headers:
        Retry-After:
          description: Seconds until the window rolls
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code]
          properties:
            code: { type: string, description: Stable machine-readable code }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              description: Optional per-code payload (e.g. Zod issues on VALIDATION)
...